rename outputGenerator -> labelGenerator
